+++ dango/testing/tests/warp.rs
@@ -3,7 +3,7 @@ use {
     dango_gateway::REVERSE_ROUTES,
     dango_testing::{
         HyperlaneTestSuite, TestOption,
-        constants::{mock_ethereum, mock_solana},
+        constants::{MOCK_HYPERLANE_LOCAL_DOMAIN, mock_ethereum, mock_solana},
         setup_test, setup_test_with_indexer,
     },
     dango_types::{
@@ -15,7 +15,6 @@ use {
     grug_math::{NumberConst, Uint128},
     grug_testing::BalanceChange,
     grug_types::{Addr, Addressable, HashExt, QuerierExt, ResultExt, StdError, btree_map, coins},
-    hyperlane_testing::constants::MOCK_HYPERLANE_LOCAL_DOMAIN,
     hyperlane_types::{
         Addr32, IncrementalMerkleTree, addr32,
         mailbox::{self, MAILBOX_VERSION, Message},
