--- hyperlane/testing/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-authors       = { workspace = true }
-categories    = { workspace = true }
-documentation = { workspace = true }
-edition       = { workspace = true }
-license       = { workspace = true }
-name          = "hyperlane-testing"
-repository    = { workspace = true }
-rust-version  = { workspace = true }
-version       = { workspace = true }
-
-[dependencies]
-eth-utils       = { workspace = true }
-grug-types      = { workspace = true }
-hex-literal     = { workspace = true }
-hyperlane-types = { workspace = true }
-k256            = { workspace = true }
-rand            = { workspace = true }
