--- hyperlane/testing/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod constants;
-mod mock_validator_set;
-
-pub use mock_validator_set::*;
